backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
gmqtt==0.6.12
sse-starlette==1.8.2
httpx==0.26.0
python-dotenv==1.0.0
bcrypt==4.1.2
aiosmtplib==3.0.1
email-validator==2.1.0.post1
